Update 2 files filter for the table
